face_recognition_server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.1.0
dlib==19.24.2
cmake==3.27.7
